major usage changes
Renamed ConsoleDisplay to ConsoleClient
Renamed GraphicalDisplay to GraphicalClient
entirely scrapped the current system of using the code editor to do the coding in favor of a console-centric variant that uses actual files for assembly, object, and executable files. This'll be much closer to the desired realism and gives an actual purpose to keeping object files lying around.
Added console commands for controlling assembly, linking, and execution. This also means that CSX64 can now be used alongside makefile or other console automation utilities for development. woot.
Because of these changes, removed all the editing/compiling features from ProcessorView. It now acts as I initially intended it to: just viewing processor internals.
This also removes a lot of the annoying bits of code from the ProcessorView class, such as state information controlling the clickability of the compile/execution utility buttons.
Because control of client forms (ProcessorView, ConsoleClient, GraphicalClient, etc.) is being migrated to external control from main, began creating access code for things that used to be publicly controlled or controlled by ProcessorView.
Also made constructors for client forms take their processor object via constructor (this way they're not instantiating their own). This makes more sense now, as they're just executing the code and providing a user interface. They shouldn't also be creating the executer.
Reordered public flags display order in the ProcessorView since it was changed in the binary specification.
Added an extension method to convert a string into a stream for potentially using stream objects for assembly/linking in the future.
Optimized Expr class to ensure that it doesn't have any unnecessary data now that it's being de/serialized all over the place. Migrated Expr class data caching into a separate function to do this elegantly.