ObjectFile.Clean property
added a public boolean property named Clean to object files that keeps track of if the object file is untampered.
you can only save clean object files.
an object file loaded from a file is clean only if all the data was read successfully.
an object file resulting from assembling an assembly file is only clean if it succeeds.
only clean object files can be passed to the linker.
if an object file is modified by the linker it is made dirty.

this property makes it possible to safely reuse object files in memory after a linking operation (e.g. perhaps you need a large library of object files loaded from the disk in case the user references them in their program. now, if the user doesn't use some of them, they will remain clean and can simply be reused as-is rather than have to reload them from the disk)