reorganization + optimization
renamed csx64 namespace to CSX64
renamed CSX64 class to Computer
split Computer class into several files in src/CSX64
moved all assembly-related things to their own static class: Assembly (since it's technically not related to the Computer class itself)

brought nested classes in Computer and Assembly to namespace scope (except private nested classes)

as there are assumptions elsewhere that addresses are "really" 32-bits (due to using List<byte> and byte[] in various places), HoleData was changed to use UInt32 for addresses.
HoleData was also changed to use byte for hole size (it'll only ever be in the range 0-3 anyway. was only UInt64 to go along with the signature for the various functions that use size codes)