file size optimization + bug fixes
fixed a bug where ObjectFile.ReadFrom() was not checking to make sure it got the full contents of the data segment.
now throws FormatException if it can't read the full contents.

changed local label names to use a shorter form, with an L instead of __local_, a hexadecimal index instead of the name of the last static label, and only the low-order 4 digits of the time stamp rather than the full 16.
as a large number of labels are expected to be local, this could potentially reduce memory/disk usage appreciably.

renamed IsValidLabel() to IsValidName() to reflect that it's also used for symbols in general, not just labels.
renamed MutateLabel() to MutateName()

added a requirement that symbol names may not exceed 255 characters in length
added this to all known entrance points into symbols table, but added a sanity check in VerifyIntegrity() just to be sure.

fixed a bug where an empty label name would crash upon reaching the insertion code

added extension methods for BinaryWriter/BinaryReader in Utility to read/write ASCII C-style strings.
this was done in an attempt to reduce object file size. Unfortunately (or fortunately) the default reader/writer implementation is apparently smart enough to not use a full 4-byte length prefix if it's not needed.

fixed a bug where console execution mode would fail if 1 path was given (was supposed to be if no paths were given).

added EndOfStreamException to catches in Program.LoadObjectFile() for handling StreamReader-specific errors.
added FormatException to catch error thrown in ObjectFile.ReadFrom() if the entire data segment cannot be read.