$ and $$ assembly macros
migrated logic for getting label offset into AssembleArgs.current_line_pos. this way it makes it more clear that the value is not going to change except for at the beginning of a line (a tad safer). additionally, this allows for $ macro to be implemented easily.

added the $ and $$ macros to TryParseImm().
$ expands to an expression that points to the current line (including #base offset).
$$ expands to the name of the last static label (i.e. the beginning of the local scope).