asm cleanup + new jmp-ish format
redefined all assembly instructions in terms of helper functions in AssembleArgs class.
tidied up assembly instruction format code (e.g. named things that had nondescript placeholder names).
did the same for operator definitions in Computer.Tick() that couln't fit in one line.

changed all Jcc (and JMP) to use the IMMRM format so that addresses don't need to be a full address expression (i.e. []).
jumping to an imm or a register works the same way (but now conforms to NASM syntax)
jumping to a memory value is visually the same as all jump-ish instructions before this update (so this breaks all existing code that used jump instructions), but the value is now dereferenced, whereas before it was as if using LEA.

renamed shifts and rotates to conform to their x86_64 names (for compatability).
renamed LA to LEA
renamed UX to ZX to fit "zero extend" name (was initially SX to line up with unary MUL/

made LOOP not take a register or size argument. now affects only $0.
binary format is now identical to jump-ish variants.