memory segments update
added AsmSegment enum to represent what segment something belongs to.
created separate holes and raw binary data for the text and data segments.
made labels use "#seg" as a base offset where seg is the segment name in which they were defined (no more #base).
updated ptrdiff logic to still work for all these new segment differences (must be in same segment to work).
added a bss segment that just holds a length (will be zeroed upon program initialization).

changed Computer.Initialize() to now require 2 8-byte values at the start of the file (length of text segment and length of bss segment).
length of text segment is used to initialize a new property Computer.ReadonlyBarrier (not yet implemented but will make writing to addresses less than that an error).
length of bss segment is added to the size of the memory block allocated for execution (this segment is zeroed during initialization).

changed up a bunch of stuff in the linker to make all of this work.
changed linker predefined-symbol definition checking to use the list of ignored tokens for verify expression.

changed it so there can only be one label max per line (this is in lieu of replacing CSX64's def directive with NASM's equ semi-instruction).

changed AssembleArgs.AppendVal() to TryAppendVal().
it now takes into account what segment we're in when attempting to write a value.
TryAppendExpr() now does the same (adding holes to their specific segment's list of holes).

created a TryParseOPSizecode() which uses instant prefix imm format.
the custom size parameter logic in SplitLine() is now replaced with this (was entirely redunant).

added SEGMENT and SECTION directives (synonyms) to control what segment we're working in: text, .data, or .bss (not case sensitive).

changed last_static_label setting in Assemble() to look for a '.' anywhere in the token (in lieu of changing to a NASM form of local labels).

made a Dump() function in Utility.cs to write a memory dump representation to the console (was used for debugging the linker).