flags update + bug fixes
updated the flags register to be x86_64 compliant.
most new flags are still not in use.
reserved flag 1 set on program initialization.
public flags are now the low 32 bits.
upper 32 (reserved in x86_64 standard) will be for CSX64-specific utilities.
renamed all the flags to end in F (CF, AF, OF, etc.).
these flags are now controlled internally by a set of masks (which are also made public for convenience in external code).

went through every instruction and recoded all of their flags to be x86_64 compatible (smoother transition from CSX64 to real assembly).
flags that are marked in the standard as undefined are loaded with a random 1 or 0 by that instructon.
floating point operations are still non-standard (don't even use FPU registers), but this is for user convenience.
only FCMP and FCMPZ affect flags now.
renamed SWAP to x86 XCHG.
removed ABS (wasn't in the x86 standard).

fixed a bug where ANDN wasn't using the settings byte to extract the size code (was using the unary value itself).

now $14 and $15 are both initialized to the top of the stack segment on program initialization.
$14 will now be treated as RBP if the user wants this behavior.
reversed the push order for command line arguments to be RTL.

made it so writing executable code to any segment except TEXT an assemble-time error.

changed __prog_end__ macro to __heap__ (makes more sense with the new segmentation).

updated the documentation file with most of the new stuff.