added FXTRACT and FSCALE
added a Utility function to extract the exponent and significand of a 64-bit floating-point value.
added a Utility function to reassemble an exponent and significand into its respective value.
did testing on these to make sure they work with a wide range of (finite) values including double.Epsilon.
the reassembled value was always identical to the original (error of zero on e17 output format).

added FXTRACT and FSCALE.

added a StackBarrier property to Computer class (if stack operations enter addresses below this it triggers ErrorCode.StackOverflow).