new and improved address format
removed AddressPrefixes array (using a new, clevererer system).
added a utility extension method string.StartsWithToken(string) - returns true if the the strings are equal or if it starts with the value and is followed by white space.
this utility function is now used to handle the explicit memory operand size cases so that "dword      ptr" is still legal.

updated the memory address format.
now only allows multipliers of 1, 2, 4, or 8, and only on a single register.
now does not allow any registers to be pre-negated (didn't notice the source i'd used had listed that as illegal).
now features a sizecode field for the address expression itself (separate from the one for the memory operand).
this allows you to put in different-width imm bases and use non-64-bit registers for addressing.
this is good, as i definitely want to promote compatibility with 32-bit assembly (also supports 16-bit, but that's not nearly as common).
8-bit addressing is not allowed (in compliance with the standard, plus 256 bytes is stupidly small).

added a utility function to extract 2 distinct powers of 2 from an integer (non-zero).
this is planned to be used later on to facilitate conversion of e.g. [edi*5] into [edi*4+edi] so that it can compile.