vec move bug fix
fixed a bug where the interpreter wasn't checking the zmask flag in mode 2.
fixed a few bugs in the assembler for the vec move family where it was refering to the wrong parsing location or using a memory size without checking to make sure it was explicit first.

updated help message to give a url for submitting bugs.