ALIGN directive + manual update
added ALIGN directive to align the next write in that segment.
added utility functions to align an address and to align a List<byte> for the next append operation.

ObjectFile now has 4 fields representing the alignment of each segment (that segment in the executable must have that alignment).
added TryAlign() function for AssembleArgs class to handle alignment on current segment.
updated the linker to simply align each segment before merging a new file's segment and to align each result segment to the next segment's alignment.

updated the manual with VPU info, but not VPU instructions.