linker alignment bug fix
fixed a bug where the linker's result segment alignment phase wasn't accounting for its own starting address in the result (was only aligning itself for the next segment to be added to it, with no regard for who came after that).

changed alignment utility functions:
AlignOffset() gets the offset needed to align an address.
Pad() adds a number of (zeroed) bytes to the end of the array.
both forms of the Align() function the same but use the above two new functions.