FSTCW/FLDCW + FPU_PC impl
fixed a bug where the binary format was using the wrong size code in the (mem, imm) case.

added the M format (only one operand, which is memory).
added the FSTCW and FLDCW instructions.

implemented FPU_PC (precision control)

renamed the binary format interpreter (removed "New" from the end that was added during a refactor pass).

with the addition of an x87 opcode, a lot of the opcodes in the manual are wrong now (not super pressing).