cleaning up stuff
changed compile settings to build to "release" and "debug" instead of to "bin/Release" and "bin/Debug".
updated the README with the new information (only locations change).

removed all the (extremely old) example code, as it used the old syntax and none of it would compile.
preserved "graphics test.asm" as "graphics_test.asm" and updated it to work with the new syntax.

added a comment to update.sh to explicitly say what it does.