stdlib: exit/atexit
added exit() and atexit() stdlib functions.
changed the _start file to call exit() instead of directly calling sys_exit.

passing a null pointer to realloc is now equivalent to calling malloc with the provided size.

fixed a bug in realloc where it wasn't deallocating the old array in the case where it couldn't be performed in-place.