asm optimization
moved code for generating segment offsets into a pair of dictionaries (was previously using c#'s implicit ToString() for the enum).
this means the segment names are now easily changed via the dictionary.
segment names were shortened to #x where x is a 1-letter id for the segment (offsets are lowercase, origins are uppercase).

changed loop-based segment merging in the linker to use AddRange() instead (i thought it'd be slower due to having to go through the IEnumerable interface, but apparently it's actually very well-optimized for that).