bug fixed + optimization
fixed a bug in the assembler/interpreter where it was writing/reading the regs byte of an address expression even when it wasn't supposed to (was using 0x0c mask instead of 0x03 - might be from a past update).
wasn't causing problems because the interpreter/assembler were putting/using the "present" flags correctly, so it'd only cause an issue if the byte wasn't there, which would map to sizecode 0, which isn't going to happen anyway since 8-bit addressing is explicitly forbidden.

did a few code analysis runs - FetchBinaryFormat() appears to be a big slowdown (because of how common it is).
optimized the function a bit:
required assignments (due to out params) are only done when absolutely necessary (and often in the fail cases).
now if get_a is false, a is undefined (instead of being zeroed) - this sort of makes more sense anyway.