added LOCK + REP RET
merged updates from the C++ impl, which added LOCK and the REP RET family of pseudo-instructions.

to get the same routing behavior for LOCK as in the C++ impl, the LOCK router casing had to be moved to the Assemble() function itself.
this is because the C# impl uses cases for instruction routing instead of hash maps like the C++ impl.